6/23 Update
A heavily WIP refactor of the entire skill system. It should allow for more than 2 skills to be assigned to the character or enemies.

This will essentially completely rewrite the code for the combat process as well.

Skills are now classes or subclasses for each one and no longer dictionaries. This should be faster in the long run.