6/24 Update
Refactor is still in major progress.

Changed all of the skills to class based objects instead of dictionaries. This should be more computationally efficient.

Re-wrote some of the code for the function that allows you to select your skills based on your class. Turned code that was 200 lines long into code that is 30 lines long. Much more efficient and clean.