6/28 Update
Finished the code for the skill selection menu. Now the player can choose skills that apply to his role and cannot have a repeat skill. This code worked before, but this code is roughly 1/5th the size and complexity for the same result. Reworked some of the combat code as well to reflect the new class changes.

Still needs a lot of work to bring it back to it's previous functionality with the 2 skill system. But now the code allows for unlimited skills and potential  for an XP system where the player adds skills as the game progresses.