7/1 Update
Rewrote the attack system and the system that checks for active buffs vs how many turns have passed. It's a lot more simple and it's not a separate function for both the player and the enemy. There's just a single check for any buffs what-so-ever. The attack action is now utilized by both the player and the enemy as well.

Added some legibility to the user interface so that it's more readable.

Still working out some bugs before pushing to the main development branch.